+++ src/lua/config.cpp
@@ -459,6 +459,14 @@ l_nnwm_focus_dir(lua_State *L)
 }
 
 static int
+l_nnwm_move_dir(lua_State *L)
+{
+    const char *dir = luaL_checkstring(L, 1);
+    nnwm::focus::move_dir(get_server(L), dir);
+    return 0;
+}
+
+static int
 l_nnwm_focus_mode_toggle(lua_State *L)
 {
     nnwm::focus::mode_toggle(get_server(L));
@@ -1324,6 +1332,7 @@ static const struct luaL_Reg nnwm_funcs[] = {
     {"focus_prev_float", l_nnwm_focus_prev_float},
     {"focus_mode_toggle", l_nnwm_focus_mode_toggle},
     {"focus_dir", l_nnwm_focus_dir},
+    {"move_dir", l_nnwm_move_dir},
     {"swap_left", l_nnwm_swap_left},
     {"swap_right", l_nnwm_swap_right},
     {"swap_next", l_nnwm_swap_next},